requirements.txt
flask
flask-session
google-generativeai
gunicorn
python-dotenv
requests